requirements.txt
pytest~=8.3.5
requests~=2.32.3
faker~=37.1.0
allure-python-commons~=2.14.0